==> notebooks/2-simulate-data.Rmd <==
---
title: "Simulation population data and selection mechanism"
output: html_notebook
---


# Simulation study 1 -- non-parametric model for multinomial case 

x1=1, x2=1, y1, y2, y3 - 0.6, 0.3, 0.1 
x1=1, x2=2, y1, y2, y3 - 0.3, 0.4, 0.3
x1=2, x2=1, y1, y2, y3 - 0.1, 0.3, 0.6
x1=2, x2=2, y1, y2, y3 - 0.2, 0.4, 0.4

```{r}
set.seed(123)
rep_x <- as.integer(500*c(50, 100, 150, 200))
df <- data.frame(x = rep(c(11,12, 21, 22), rep_x))
df$x1 <- ifelse(substr(df$x,1,1) == 1,  1, 2)
df$x2 <- ifelse(substr(df$x,2,2) == 1,  1, 2)
df$y <- NA
df$y <- ifelse(df$x == 11, sample(x = c(1,2,3), size = rep_x[1], prob = c(0.6 ,0.3, 0.1), replace = T), df$y)
df$y <- ifelse(df$x == 12, sample(x = c(1,2,3), size = rep_x[2], prob = c(0.3, 0.4, 0.3), replace = T), df$y)
df$y <- ifelse(df$x == 21, sample(x = c(1,2,3), size = rep_x[3], prob = c(0.1, 0.3, 0.6), replace = T), df$y)
df$y <- ifelse(df$x == 22, sample(x = c(1,2,3), size = rep_x[4], prob = c(0.2, 0.4, 0.4), replace = T), df$y)

prop.table(table(df$y))

## nmar (x1 and y)
#eta <- (df$y == 2)
eta <- -0.8*(df$y == 2) - 0.4*(df$x1 == 2)
rho <- exp(eta) / (1 + exp(eta))
df$probs_nmar <- rho

## mar (only on x1, x2)
eta <- 1 + (df$x1 == 1)
rho <- exp(eta) / (1 + exp(eta))
df$probs_mar <- rho

saveRDS(object = df, file = "../data/df-sim-nonpar.rds")
```

# Simulation study 2 -- parametric model for multinomial case


```{r}
set.seed(123)
rep_x <- as.integer(500*c(50, 100, 150, 200))
df <- data.frame(x = rep(c(11,12, 21, 22), rep_x))
df$x1 <- ifelse(substr(df$x,1,1) == 1,  1, 0)
df$x2 <- ifelse(substr(df$x,2,2) == 1,  1, 0)
df$eta1 <- 1
df$eta2 <- exp(1 - 1*df$x1 + 1*df$x2)
df$eta3 <- exp(1 + 2*df$x2 - 2*df$x2)
df$pr1 <- df$eta1/(df$eta1+df$eta2+df$eta3)
df$pr2 <- df$eta2/(df$eta1+df$eta2+df$eta3)
df$pr3 <- df$eta3/(df$eta1+df$eta2+df$eta3)
df$y <- apply(df, MARGIN = 1, FUN = function(x) sample(x = 1:3, size = 1, replace = T, prob = x[c('pr1','pr2','pr3')]))
prop.table(table(df$y))
```


# Simulation study 3 -- using real data from cbop

```{r}
cbop <- readRDS(file = "../data/df-cbop-2018Q1.rds")
cbop$y <-factor(cbop$kod_rodz_zatrud, 
                c("Umowa o pracę na okres próbny", "Umowa o pracę na czas określony",
                  "Umowa o pracę na czas nieokreślony"),
                1:3)
cbop$id <- 1:nrow(cbop)
cbop <- cbop[,c("id", "nr_oferty", "zawod1", "kod_rodz_zatrud", "y", "nace")]

# Missing data mechanisms

## 1 and 3 category are less likely to be observed
eta_13 <- 1.5*(cbop$y == 1) + 1.5*(cbop$y == 3)
rho_13 <- exp(eta_13) / (1 + exp(eta_13))
cbop$probs_nmar_13 <- rho_13

eta_2 <- 1 -0.8*(cbop$y == 2)
rho_2 <- exp(eta_2) / (1 + exp(eta_2))
cbop$probs_nmar_2 <- rho_2

#cbop$flag_13 <- rbinom(n = nrow(cbop), size = 1, prob = cbop$probs_nmar_13)
#cbop$flag_2 <- rbinom(n = nrow(cbop), size = 1, prob = cbop$probs_nmar_2)

#table(cbop$flag_13)
#table(cbop$flag_2)

#prop.table(table(cbop$y[cbop$flag_13==1]))
#prop.table(table(cbop$y[cbop$flag_2==1]))
#prop.table(table(cbop$y))
saveRDS(object = cbop, file = "../data/df-sim-cbop.rds")

```
