==> notebooks/3-cbop-vis-rap.Rmd <==
---
title: "R Notebook"
output: html_notebook
---

```{r}
library(tidyverse)
library(maxLik)
library(scales)
library(ca)
library(vcd)
theme_set(theme_bw())
```

```{r}
cbop <- readRDS(file = "../data/df-cbop-2018Q1.rds")
totals <- readRDS(file = "../data/df-demand-totals.rds")
```

Compare distributions

-- occupancy

```{r}
cbop %>%
  filter(przez_prac) %>%
  count(occup = zawod1, wt = liczba_miejsc_og, name = "cbop")  %>%
  left_join(totals %>% count(occup, wt = jvs, name = "jvs")) -> occup_vals


occup_vals %>%
  mutate(p = cbop / jvs*100)

occup_vals %>%
  gather(stats, vals, -occup) %>%
  mutate(occup = as.factor(occup)) %>%
  ggplot(data = ., aes(x = occup, y = vals, fill = stats)) +
  geom_col(position = "dodge", color = "black") + 
  labs(x = "Occupation", y = "Number of vacancies", fill = "Source") +
  scale_fill_brewer(type = "qual", palette = "Set1")

occup_vals %>%
  gather(stats, vals, -occup) %>%
  mutate(occup = as.factor(occup)) %>%
  add_count(stats, wt = vals, name = "total") %>%
  mutate(p = vals/total) %>%
  ggplot(data = ., aes(x = occup, y = p, fill = stats)) +
  geom_col(position = "dodge", color = "black") + 
  labs(x = "Occupation", y = "Shares", fill = "Source") +
  scale_fill_brewer(type = "qual", palette = "Set1") +
  scale_y_continuous(labels = percent_format())

```

-- nace

```{r}
cbop %>%
  filter(przez_prac) %>%
  count(nace, wt = liczba_miejsc_og, name = "cbop")  %>%
  left_join(totals %>% count(nace, wt = jvs, name = "jvs")) -> nace_vals

nace_vals %>%
  mutate(p = cbop / jvs*100)

nace_vals %>%
  gather(stats, vals, -nace) %>%
  mutate(nace = as.factor(nace)) %>%
  ggplot(data = ., aes(x = nace, y = vals, fill = stats)) +
  geom_col(position = "dodge", color = "black") + 
  labs(x = "NACE", y = "Number of vacancies", fill = "Source") +
  scale_fill_brewer(type = "qual", palette = "Set1")

nace_vals %>%
  gather(stats, vals, -nace) %>%
  mutate(nace = as.factor(nace)) %>%
  add_count(stats, wt = vals, name = "total") %>%
  mutate(p = vals/total) %>%
  ggplot(data = ., aes(x = nace, y = p, fill = stats)) +
  geom_col(position = "dodge", color = "black") + 
  labs(x = "NACE", y = "Shares", fill = "Source") +
  scale_fill_brewer(type = "qual", palette = "Set1") +
  scale_y_continuous(labels = percent_format())
```

Correlation with target variable

```{r}
tab_nace <- xtabs(liczba_miejsc_og ~ nace + kod_rodz_zatrud, data = cbop)
tab_occup <- xtabs(liczba_miejsc_og ~ zawod1 + kod_rodz_zatrud, data = cbop)
tab_woj <- xtabs(liczba_miejsc_og ~ kod_woj + kod_rodz_zatrud, data = cbop)

tab_nace_zmiany <- xtabs(liczba_miejsc_og ~ nace + zmiany, data = cbop)
tab_occup_zmiany <- xtabs(liczba_miejsc_og ~ zawod1 + zmiany, data = cbop)
tab_woj_zmiany <- xtabs(liczba_miejsc_og ~ kod_woj + zmiany, data = cbop)


```

```{r}
cat("==============================\n")
assocstats(tab_nace)
cat("==============================\n")
assocstats(tab_occup)
cat("==============================\n")
assocstats(tab_woj)
cat("==============================\n")
assocstats(tab_nace_zmiany)
cat("==============================\n")
assocstats(tab_occup_zmiany)
cat("==============================\n")
assocstats(tab_woj_zmiany)
cat("==============================\n")
```

```{r}
ca(tab_nace) %>% plot()
ca(tab_occup) %>% plot()
ca(tab_woj) %>% plot()

ca(tab_nace_zmiany) %>% plot()
ca(tab_occup_zmiany) %>% plot()
ca(tab_woj_zmiany) %>% plot()

```
